implemented ad delete
