+++ internal/database/ad_change_storage.go
@@ -1,0 +1,122 @@
+package database
+
+import (
+	"fmt"
+	"github.com/sergeychur/give_it_away/internal/models"
+	"gopkg.in/jackc/pgx.v2"
+)
+
+const (
+	// create ad query
+	CreateAd = "INSERT INTO ad (author_id, header, text, region, district, is_auction, feedback_type, category%s)" +
+		" VALUES($1, $2, $3, $4, $5, $6, $7, $8%s) RETURNING ad_id"
+	ExtraField = ", extra_field"
+	GeoPosition = ", geo_position, lat, long"
+	Blank = ""
+	NoExtraFieldNoGeoPosition = ""
+	NoExtraFieldGeoPosition = ", ST_POINT($9, $10), $9, $10"
+	ExtraFieldNoGeoPosition = ", $9"
+	ExtraFieldGeoPosition = ", $9, ST_POINT($10, $11), $10, $11"
+
+	// add photo to ad query
+	checkAdExist = "SELECT author_id FROM ad WHERE ad_id = $1"
+	AddPhotoToAd = "INSERT INTO ad_photos (ad_id, photo_url) VALUES ($1, $2)"
+
+	// deleteAd query
+	deleteAd = "DELETE FROM ad WHERE ad_id = $1"
+)
+
+func (db *DB) CreateAd(ad models.Ad) (int, models.AdCreationResult) {
+	query := ""
+	var err error
+	err = nil
+	res := models.AdCreationResult{}
+	sign := 0
+	if ad.FeedbackType == Other {
+		sign = 10
+	}
+	if ad.GeoPosition != nil {
+		sign += 1
+	}
+	switch sign {
+	case 0:
+		query = fmt.Sprintf(CreateAd, Blank, NoExtraFieldNoGeoPosition)
+		err = db.db.QueryRow(query, ad.AuthorId, ad.Header, ad.Text, ad.Region, ad.District, ad.IsAuction,
+			ad.FeedbackType, ad.Category).Scan(&res.AdId)
+	case 1:
+		query = fmt.Sprintf(CreateAd, GeoPosition, NoExtraFieldGeoPosition)
+		err = db.db.QueryRow(query, ad.AuthorId, ad.Header, ad.Text, ad.Region, ad.District, ad.IsAuction,
+			ad.FeedbackType, ad.Category, ad.GeoPosition.Latitude, ad.GeoPosition.Longitude).Scan(&res.AdId)
+	case 10:
+		query = fmt.Sprintf(CreateAd, ExtraField, ExtraFieldNoGeoPosition)
+		err = db.db.QueryRow(query, ad.AuthorId, ad.Header, ad.Text, ad.Region, ad.District, ad.IsAuction,
+			ad.FeedbackType, ad.Category, ad.ExtraField).Scan(&res.AdId)
+	case 11:
+		query = fmt.Sprintf(CreateAd, ExtraField + GeoPosition, ExtraFieldGeoPosition)
+		err = db.db.QueryRow(query, ad.AuthorId, ad.Header, ad.Text, ad.Region, ad.District, ad.IsAuction,
+			ad.FeedbackType, ad.Category, ad.ExtraField, ad.GeoPosition.Latitude, ad.GeoPosition.Longitude).Scan(&res.AdId)
+	}
+	if err != nil {
+		return DB_ERROR, res
+	}
+	return CREATED, res
+}
+
+func (db *DB) AddPhotoToAd(pathToPhoto string, adId int, userId int) int {
+	tx, err := db.StartTransaction()
+	if err != nil {
+		return DB_ERROR
+	}
+	defer func() {
+		_ = tx.Rollback()
+	}()
+	authorId :=0
+	err = tx.QueryRow(checkAdExist, adId).Scan(&authorId)
+	if err != nil {
+		return DB_ERROR
+	}
+	// TODO: uncomment when we can take userId from cookies
+	/*if authorId != userId {
+		return CONFLICT
+	}*/
+	_, err = tx.Exec(AddPhotoToAd, adId, pathToPhoto)
+	if err != nil {
+		return DB_ERROR
+	}
+	err = tx.Commit()
+	if err != nil {
+		return DB_ERROR
+	}
+	return FOUND
+}
+
+func (db *DB) DeleteAd(adId int, userId int) int {
+	tx, err := db.StartTransaction()
+	if err != nil {
+		return DB_ERROR
+	}
+	defer func() {
+		_ = tx.Rollback()
+	}()
+	authorId := 0
+	err = tx.QueryRow(checkAdExist, adId).Scan(&authorId)
+	if err == pgx.ErrNoRows {
+		return EMPTY_RESULT
+	}
+	if err != nil {
+		return DB_ERROR
+	}
+	// TODO: uncomment when we can take userId from cookies
+	/*if authorId != userId {
+		return CONFLICT
+	}*/
+	_, err = tx.Exec(deleteAd, adId)
+	if err != nil {
+		return DB_ERROR
+	}
+	err = tx.Commit()
+	if err != nil {
+		return DB_ERROR
+	}
+	return FOUND
+}
